Methods restructured, body updates DOM
