+++ src/Сomponents/Body/Body.js
@@ -1,15 +1,24 @@
 import React, { Component } from 'react';
 import styles from './Body.module.css';
 import globalStyles from '../../CSS/global.css';
+import { Form } from '../Form';
 
-import { PopupWelcome } from '../PopupWelcome/';
-import { Button } from '../Button';
-import { Text } from '../Text'; 
+export class Body extends React.Component {
+  constructor(props) {
+    super(props);
+    this.handleCutUpArray = this.handleCutUpArray.bind(this);
+    this.state = {cutUps: []};
+  }
 
-export function Body() {
-  return(
-    <main>
-      <Text />
-    </main>
-  );
+  handleCutUpArray = cutUpArray => this.setState({cutUps: cutUpArray});
+
+  render() {
+    return(
+      <main>
+        <Form onSubmit={this.handleCutUpArray}/>
+        {this.state.cutUps}
+      </main>
+    );
+  }
 }
+
