+++ src/Сomponents/Body/Body.module.css
@@ -1,0 +1,3 @@
+body {
+  background: #cccccc;
+}
