+++ src/Сomponents/Form/index.js
@@ -1,0 +1,1 @@
+export { Form } from './Form.js'
